requirements.txt
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg2
